% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importAttributes.R
\name{importAttributes}
\alias{importAttributes}
\title{Imports/combines CAMELS attribute files (see N. Addor et al., 2017)}
\usage{
importAttributes(folder, attrSubset = NULL, returnList = FALSE)
}
\arguments{
\item{folder}{path/name of folder with CAMELS attribute text files}

\item{attrSubset}{string vector of attribute file types to import (e.g. 'soil')}

\item{returnList}{Bool, return list of data by file instead of data frame?}
}
\value{
dataframe or list of dataframes
}
\description{
Reads in all CAMELS attribute text files and returns a data frame (or list)
of the data
}
\examples{
basinPar <- importAttributes('camels_attributes_v2.0')
}
\author{
Leland Scantlebury
}
