% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importHRUs.R
\name{importHRUs}
\alias{importHRUs}
\title{Import CAMELS HRU Daymet forcings}
\usage{
importHRUs(folder, subset = NULL, verbose = TRUE, ...)
}
\arguments{
\item{folder}{name/path HRU forcing folder (e.g., 'hru_forcing')}

\item{subset}{string vector of basin ids to import (optional, otherwise imports all)}

\item{verbose}{T/F to determine whether to print importing status}

\item{...}{optional arguments passed to read.HRUForcing()}
}
\value{
list of lists (of lists)
}
\description{
Reads all (unless subset passed) of the HRU forcing files. Returns list of
lists - outer most list index is the basin id, which give access to a list
indexed by the basin's HRU ids. The HRU list contains four slots:
 - data (dataframe of HRU forcing data)
 - gaugeLatitude
 - HRUElevation
 - HRUArea
}
\examples{
basinSubset <- c('01013500','10259200','14137000',
                 '08198500','03592718','03604000')
HRUForcings <- importHRUs(folder = 'basin_dataset_public_v1p2/hru_forcing/daymet/',
                          subset = basinSubset )
}
\author{
Leland Scantlebury
}
