% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importStreamflows.R
\name{importStreamflows}
\alias{importStreamflows}
\title{Import USGS Streamflow files}
\usage{
importStreamflows(folder, subset = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{folder}{name/path streamflow folder (e.g., 'usgs_streamflow')}

\item{subset}{string vector of basin ids to import (optional, otherwise imports all)}

\item{verbose}{T/F to determine whether to print importing status}

\item{...}{optional arguments passed to read.streamflow()}
}
\value{
list of dataframes
}
\description{
Reads in all streamflow files within folders (unless subset is specified)
and returns a list of dataframes indexed by basin id of the stream gauge
data.
}
\examples{
basinSubset <- c('01013500','10259200','14137000',
                 '08198500','03592718','03604000')
streamflows <- importStreamflows('basin_dataset_public_v1p2/usgs_streamflow',
                                 subset = basinSubset)
}
\author{
Leland Scantlebury
}
