% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.basinForcing.R
\name{read.basinForcing}
\alias{read.basinForcing}
\title{Basin Forcing File Reader}
\usage{
read.basinForcing(filename, altheader = NULL, dropDateInts = TRUE)
}
\arguments{
\item{filename}{filename/path of basin forcing text file}

\item{altheader}{vector string to be used as column names (optional)}

\item{dropDateInts}{T/F to drop original integer date columns after
conversion to date format (default T)}
}
\value{
list
}
\description{
Returns named list with four items:
 - data (dataframe of forcing data)
 - gaugeLatitude
 - gaugeElevation
 - basinArea
}
\examples{
folder <- 'basin_dataset_public_v1p2/basin_mean_forcing/maurer/11/'
filename <- paste0(folder, '07197000_lump_maurer_forcing_leap.txt')
bforce <- read.basinForcing(filename)
}
\author{
Leland Scantlebury
}
