% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.basinInfo.R
\name{read.basinInfo}
\alias{read.basinInfo}
\title{Read CAMELS basin Physical Characterists File}
\usage{
read.basinInfo(filename, altheader = FALSE)
}
\arguments{
\item{filename}{filename/path of basin characterists file (e.g. 'basin_physical_characteristics.txt')}

\item{altheader}{vector string of custom column names to be used (optional)}
}
\value{
dataframe
}
\description{
Read CAMELS basin Physical Characterists File
}
\examples{
folder <- 'basin_dataset_public_v1p2/basin_metadata/'
filename <- paste0(folder, 'basin_physical_characteristics.txt')
basinChar <- read.BasinInfo(filename)
}
\author{
Leland Scantlebury
}
