% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.camelsAttr.R
\name{read.camelsAttr}
\alias{read.camelsAttr}
\title{Read CAMELS attribute file (see N. Addor et al., 2017)}
\usage{
read.camelsAttr(filename, altheader = NULL, subset = NULL)
}
\arguments{
\item{filename}{filename/path of basin forcing text file}

\item{altheader}{vector string to be used as custom column names (optional)}

\item{subset}{list of basins to subset and return}
}
\value{
dataframe
}
\description{
Generic text file reader for the CAMELS attribute files
}
\examples{
geol <- read.camelsAttr(filename = 'camels_geol.txt')
}
\author{
Leland Scantlebury
}
