% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.gaugeInfo.R
\name{read.gageInfo}
\alias{read.gageInfo}
\title{Read CAMELS river gage information file.}
\usage{
read.gageInfo(filename, altheader = FALSE)
}
\arguments{
\item{filename}{filepath/name of CAMELS gauge_information.txt file}

\item{altheader}{alternative header for applying to DataFrame columns}
}
\value{
DataFrame of river gauge info
}
\description{
Alternative spelling edition of read.gaugeInfo()
}
\examples{
  gages <- read.gageInfo('basin_dataset_public_v1p2/basin_metadata/gauge_information.txt')
}
\author{
Leland Scantlebury
}
\keyword{CAMELS}
\keyword{gage}
\keyword{gauge}
\keyword{river}
