% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.gaugeInfo.R
\name{read.gaugeInfo}
\alias{read.gaugeInfo}
\title{Read CAMELS river gauge information file.}
\usage{
read.gaugeInfo(filename, altheader = FALSE)
}
\arguments{
\item{filename}{filepath/name of CAMELS gauge_information.txt file}

\item{altheader}{alternative header for applying to DataFrame columns}
}
\value{
DataFrame of river gauge info
}
\description{
A file reader for the CAMELS river gauge information file. Provides
correct header (as of June 2019) and uses read.table with specific parameters
to ensure the file is read correctly.
}
\examples{
  gauges <- read.gaugeInfo('basin_dataset_public_v1p2/basin_metadata/gauge_information.txt')
}
\author{
Leland Scantlebury
}
\keyword{CAMELS}
\keyword{gage}
\keyword{gauge}
\keyword{river}
