% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.HRUForcing.R
\name{read.HRUForcing}
\alias{read.HRUForcing}
\title{Read CAMELS Basin HRU Forcing file}
\usage{
read.HRUForcing(filename, altheader = FALSE, dropDateInts = TRUE)
}
\arguments{
\item{filename}{filename/path of HRU forcing text file}

\item{altheader}{vector string to be used as column names (optional)}

\item{dropDateInts}{T/F to drop original integer date columns after
conversion to date format (default T)}
}
\value{
list
}
\description{
Returns namedlist with four items:
 - data (dataframe of forcing data)
 - gaugeLatitude
 - HRUElevation
 - HRUArea
}
\examples{
myhru <- read.HRUForcing('hru_forcing/daymet/02/01333000_hru_04535_cida_forcing_leap.txt')
}
\author{
Leland Scantlebury
}
