% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.streamflow.R
\name{read.streamflow}
\alias{read.streamflow}
\title{Read CAMELS USGS Streamflow File}
\usage{
read.streamflow(
  filename,
  altheader = FALSE,
  dropDateInts = TRUE,
  convertNA = TRUE
)
}
\arguments{
\item{filename}{filename/path of HRU forcing text file}

\item{altheader}{vector string to be used as column names (optional)}

\item{dropDateInts}{T/F to drop original integer date columns after
conversion to date format (default T)}

\item{convertNA}{T/F convert NA values (-999) to R-friendly NA}
}
\value{
dataframe
}
\description{
Read CAMELS USGS Streamflow File
}
\examples{
folder <- 'basin_dataset_public_v1p2/usgs_streamflow/01/'
filename <- paste0(folder, '01013500_streamflow_qc.txt')
mystream <- read.streamflow(filename)
}
\author{
Leland Scantlebury
}
